tools/requirements.txt
async==0.6.2
boto==2.38.0
cffi==1.0.0
cov-core==1.14.0
coverage==3.7.1
dj-database-url==0.2.2
dj-static==0.0.6
Django==1.7
django-admin-tools==0.5.2
django-filter==0.7
django-grappelli==2.5.7
django-guardian==1.1.1
django-heroku-memcacheify==0.5
django-mathfilters==0.2.1
django-nose==1.2
django-pylibmc-sasl==0.2.4
django-toolbelt==0.0.1
django-userena==1.4.0
djangorestframework==3.0.5
docopt==0.4.0
docutils==0.12
easy-thumbnails==1.3
enum34==1.0.4
gdata==2.0.18
gitdb==0.5.4
GitPython==0.3.2rc1
google-api-python-client==1.2
gunicorn==18.0
heroku==0.1.4
html2text==3.200.3
httplib2==0.8
humanize==0.5.1
idna==2.0
ipaddress==1.0.7
mailchimp==2.0.9
mandrill==1.0.57
newrelic==2.50.0.39
nose==1.3.6
nose-cov==1.6
Pillow==2.3.1
pipdeptree==0.4.2
psycopg2==2.6
pyasn1==0.1.7
pycparser==2.13
pygeoip==0.3.2
pylibmc==1.2.3
python-dateutil==1.5
python-gflags==2.0
raven==5.3.1
requests==2.7.0
simplejson==3.7.1
six==1.9.0
smmap==0.9.0
static==0.4
static3==0.6.1
stripe==1.14.0
yolk==0.4.3
